backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pillow
web3
supabase
python-jose[cryptography]
passlib[bcrypt]